Fix crypto preflight blockers before opening the PR
Address four defects found by adversarial preflight review, each of
which would have failed CI:

- CommonCrypto status comparisons: CCCrypt and CCKeyDerivationPBKDF
  return Int32 while kCCSuccess imports as Int, so the equality checks
  did not compile. Cast the constant to Int32.
- RSA fixture format: openssl pkey -outform DER emits a traditional
  PKCS#1 RSAPrivateKey on OpenSSL 3.x, but the DER reader and the spec
  require the PKCS#8 PrivateKeyInfo envelope, so the organization-key
  unwrap test would fail. Emit PKCS#8 via openssl pkcs8 -topk8 -nocrypt
  and regenerate the fixtures so the private key and wrapped org key
  stay a consistent pair.

Also release the Security framework CFError out-params on the failure
paths so they do not leak, and drop stray non-ASCII from an app-source
doc comment.

+++ VaultSquire/Providers/Vaultwarden/Crypto/VaultwardenEncString.swift
@@ -201,7 +201,8 @@ enum VaultwardenCipher {
                 }
             }
         }
-        guard status == kCCSuccess else {
+        // CCCrypt returns CCCryptorStatus (Int32); kCCSuccess imports as Int.
+        guard status == Int32(kCCSuccess) else {
             throw VaultwardenCryptoError.integrityFailure
         }
